gfx: Use strings as object keys
It occurred to me that there is no reason to refer to these objects
except by their names, so maps indexed by strings with SSO or interning
is totally superior as storage---a choice so obvious, it wasn't.

Also created a somewhat bloated new test shader that uses a handful more
features.