Fix/silence clippy errors
Considering I have never run clippy before on this repo, the number
of warnings was not too bad. Most lints were useless or pedantic or
targeted towards library developers, but the thing caught a misaligned
pointer bug, which was good.

A good schedule to fix lints might be once a month (for a solo dev).