Rename DeviceHeap -> ImageHeap
The change is relatively small, but there's not really any need for a
general-purpose heap, and this opens the opportunity to specialize the
allocator for images, e.g. it could support compacting.

As a consequence, the heap now only supports nonlinear images on device
local memory. Other cases are too specialized to require suballocation.