Remove scalar type parameter from math types
When defining the quaternion type, it became apparent that employing a
type parameter was a hindrance to actually programming with vectors and
matrices. The type parameter never served any practical purpose, so it
was essentially window dressing all along.

This commit also takes advantage of the "inherent impl" pattern to clean
up trait implementation and usage, further simplifying the library
design.